requirements.txt
Django==3.2.*
django-simpleui==2024.*
pymysql==1.1.*
cryptography==42.*
mysqlclient==2.2.*
pandas==2.0.*
scikit-learn==1.3.*
joblib==1.3.*
openai==1.30.*
requests==2.31.*
jieba==0.42.*
snownlp==0.12.*
sqlalchemy==2.0.*
djangorestframework==3.15.*
django-cors-headers==4.4.*
wordcloud==1.9.*
matplotlib==3.7.*
